GLiNER-streamlit-app/requirements.txt
streamlit
st-annotated-text
gliner